Classic 0.0.14a_08 and Infdev 20100325 (MB 1.3)

+++ src/main/java/mod/bespectacled/modernbeta/world/feature/placement/noise/NoiseBasedCountInfdev420.java
@@ -7,7 +7,7 @@ public class NoiseBasedCountInfdev420 implements NoiseBasedCount {
     private final PerlinOctaveNoise noiseSampler;
     
     public NoiseBasedCountInfdev420(Random random) {
-        this.noiseSampler = new PerlinOctaveNoise(new java.util.Random(random.nextInt()), 5, true);
+        this.noiseSampler = new PerlinOctaveNoise(new java.util.Random(random.nextLong()), 5, true);
     }
     
     public NoiseBasedCountInfdev420(PerlinOctaveNoise noiseSampler) {
